Updated SlideMenui and TurretBuilder Logic
Changed SlideMenu function from ToggleMenu to SetMenu with a boolean passed in as the parameter.
In TurretBuilder, updated the toggleMenu calls to the SetMenu calls with the appropriate boolean passed in.
Added check for the case where we click on one turret to another, the previously selected turret will call the deselect turret function.