LevelManager level start logic updates
Cleaned up naming conventions inconsistencies in code.
Created Variable levelStart that is false by default, StartWave public function can be called to start the wave whenever a condition is met. If we want to auto start level we can just tick the levelStart variable to true per level that does not need a start condition.